orchestrator_adk_example.py
"""
ADK-Based Accessibility Testing Orchestrator
Example of how to refactor using pure ADK coordination patterns
"""

import os
from typing import List, Dict, Any
from google.adk.agents import LlmAgent
from agents.keyboard_focus_agent import KeyboardFocusAgent
from agents.color_contrast_agent import ColorContrastAgent

class AccessibilityCoordinatorAgent(LlmAgent):
    """
    ADK-based coordinator agent that follows the official pattern
    Uses sub_agents for automatic coordination by the ADK engine
    """
    
    def __init__(self):
        # Create specialized sub-agents
        keyboard_agent = KeyboardFocusAgent()
        color_agent = ColorContrastAgent()
          # Initialize as LlmAgent with sub_agents
        super().__init__(
            name="accessibility_coordinator",
            model=os.getenv('CLAUDE_MODEL', 'claude-3-5-sonnet-20241022'),
            instruction="""You are an accessibility testing coordinator.
            
            Your role is to:
            1. Analyze user requests for accessibility testing
            2. Coordinate with specialized sub-agents for different tests
            3. Compile comprehensive accessibility reports
            
            Available sub-agents:
            - keyboard_focus_agent: Tests keyboard navigation and focus management
            - color_contrast_agent: Tests color contrast and visual accessibility
            
            When a user requests accessibility testing, determine which agents to invoke
            and coordinate their activities to produce a comprehensive report.""",
            description="Coordinates accessibility testing using specialized sub-agents",
            sub_agents=[keyboard_agent, color_agent]  # ADK handles this coordination
        )
    
    async def test_accessibility(self, url: str, context=None):
        """
        Test accessibility using ADK's coordination patterns
        The LLM will automatically decide which sub-agents to invoke
        """
        # ADK engine and model will handle sub-agent coordination
        # Just provide the context and let ADK route to appropriate sub-agents
        
        test_prompt = f"""
        Please perform comprehensive accessibility testing for: {url}
        
        Coordinate with the appropriate sub-agents to:
        1. Test keyboard navigation and focus management
        2. Analyze color contrast and visual accessibility  
        3. Compile results into a comprehensive report
        
        Ensure all WCAG 2.2 Level AA requirements are evaluated.
        """
        
        # In a real ADK implementation, you would invoke the agent
        # and the framework would coordinate sub-agents automatically
        return await self.invoke_async(test_prompt, context)

# Usage example following ADK patterns
def create_adk_accessibility_system():
    """Create accessibility testing system using pure ADK patterns"""
    
    # Single coordinator agent with sub-agents
    # ADK engine handles all coordination automatically
    coordinator = AccessibilityCoordinatorAgent()
    
    return coordinator

# This would replace the current AccessibilityOrchestrator class
# with a pure ADK-based approach where the LLM decides coordination 