requirements.txt
# AI Accessibility Testing Agent Dependencies
# Google ADK Python framework for multi-agent system
google-adk>=1.1.1

# A2A Protocol - install from GitHub (official PyPI package may not be available yet)
# a2a-sdk>=1.0.0
# Use: pip install git+https://github.com/google-a2a/a2a-python.git

# Core dependencies
aiohttp>=3.8.0
click>=8.0.0
colorama>=0.4.4
beautifulsoup4>=4.11.0
lxml>=4.9.0
requests>=2.28.0
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Accessibility testing
axe-selenium-python>=2.1.6

# Color contrast analysis
Pillow>=9.0.0

# Web scraping and browser automation
playwright>=1.30.0
selenium>=4.8.0

# LLM Integration
anthropic>=0.34.0
python-dotenv>=1.0.0

# Environment and Configuration
pydantic>=2.0.0
